Added automatic version update scripts incrementVersion.py and incrementRevision.py (these should be used with the bzr hook scripts pre_commit_hook.py and start_commit_hook.py, but since those go in the bzr plugins directory, I have not added them to source control).
Also added a Revision.java file to the project to compliment the Version.java file, but will not add that to source control as it will always be 1 revision behind the revision in the respository using automatic incrementation.  Revision.java is marked as ignored so it doesn't keep showing up in bzr.