Added missing cursor.close() statements
CsRestContentProvider changes:
- added missing cursor.close()
- removed old, commented-out code

CsErrorListFragment changes:
- added missing cursor.close()

CsSnapshotDetailsFragment changes:
- added missing cursor.close()

CsVmDetailsFragment changes:
- added missing cursor.close()

CsVmListFragment changes:
- added missing cursor.close()

Note that on the 2.2 emulator, a "Finalizing a Cursor that has not been deactivated or closed. "/"Application did not close the cursor or database object that was opened here" error is still showing up, but seemingly for the CursorLoader.  From what I understand from my research, CursorLoaders automatically close the cursor it was holding upon the adaptor.swapCursor() call (otherwise, what is the point of using the CursorLoader?), but for whatever reason the error is still showing up on the 2.2 emulator.  Leaving this as is for now since it doesn't seem to affect functionality.