Fixed "disappearing footers" bug for CsSnapshotListFragment
If you swipe the CsSnapshotListFragment too far off screen then get it to reload (e.g. starting from CsSnapshotListFragment, go all the way to CsAccountsFragment and then back), though the snapshot items are all reloaded correctly, the footers do not get shown.  You can get them to re-display by shifting orientation or existing and restarting the app, but not by pausing and resuming it.

Changed the setListAdapter-related code to always null-and-set even when we are re-using a previously set adapter, which seems to make the list recognize the set footers and display them.

CsVmListFragment should have the exact same problem, but since it is the middle screen right now, there doesn't seem to be a case where we can swipe far off enough to get multi-list to unload it.  Made the same fix to CsVmListFragment anyways just in case.

--
CsSnapshotListFragment/CsSnapshotListFragment
- fixed disappearing footer bug

README
- updated Cumulus 1.2 APK build number